refactor(subway): create packages to separate controllers
controllers are in different packages according to their uses. OptionMenuController now contains path controllers
