+++ src/main/java/subway/controller/menu/MainMenuController.java
@@ -1,4 +1,4 @@
-package subway.controller;
+package subway.controller.menu;
 
 import subway.view.InputView;
 import subway.view.OutputView;
