+++ src/main/java/subway/controller/menu/OptionMenuController.java
@@ -1,5 +1,7 @@
-package subway.controller;
+package subway.controller.menu;
 
+import subway.controller.path.QuickestPathController;
+import subway.controller.path.ShortestPathController;
 import subway.view.InputView;
 import subway.view.OutputView;
 
@@ -7,6 +9,8 @@ public class OptionMenuController extends MenuController {
 
     public OptionMenuController(InputView inputView) {
         super(inputView);
+        childControllers.add(new ShortestPathController(inputView));
+        childControllers.add(new QuickestPathController(inputView));
     }
 
     @Override
